# api/openapi.yaml
openapi: 3.1.0
info:
  title: ame API
  description: Adaptive Mastery Engine — quiz & exam platform with agent identity.
  license:
    name: ''
  version: 0.1.0
paths:
  /v1/admin/audit:
    get:
      tags:
      - admin
      summary: GET /v1/admin/audit — retrieve append-only audit trail logs
      operationId: list_audit_logs
      responses:
        '200':
          description: Audit trail successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuditLogsResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin required)
      security:
      - bearer: []
  /v1/admin/moderate:
    post:
      tags:
      - admin
      summary: POST /v1/admin/moderate — moderate public content by unpublishing it (setting visibility private)
      operationId: moderate_quiz
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerateBody'
        required: true
      responses:
        '204':
          description: Public content moderated and unpublished successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin required)
        '404':
          description: Content not found
      security:
      - bearer: []
  /v1/admin/users:
    get:
      tags:
      - admin
      summary: GET /v1/admin/users — list all users
      operationId: list_users
      responses:
        '200':
          description: User list successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin required)
      security:
      - bearer: []
  /v1/admin/users/{id}:
    patch:
      tags:
      - admin
      summary: PATCH /v1/admin/users/{id} — toggle user plan or disable/deactivate user
      operationId: patch_user_admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserAdminBody'
        required: true
      responses:
        '204':
          description: User successfully updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin required)
        '404':
          description: User not found
      security:
      - bearer: []
...
components:
  schemas:
...
    AuditLogEntry:
      type: object
      required:
      - id
      - action
      - metadata
      - createdAt
      properties:
        action:
          type: string
        actorUserId:
          type:
          - string
          - 'null'
          format: uuid
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        metadata: {}
        targetId:
          type:
          - string
          - 'null'
          format: uuid
        targetType:
          type:
          - string
          - 'null'
...
    ListAuditLogsResponse:
      type: object
      required:
      - logs
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
...
    ListUsersResponse:
      type: object
      required:
      - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    ModerateBody:
      type: object
      required:
      - quizId
      properties:
        quizId:
          type: string
          format: uuid
...
    PatchUserAdminBody:
      type: object
      properties:
        disabled:
          type:
          - boolean
          - 'null'
        plan:
          type:
          - string
          - 'null'
        role:
          type:
          - string
          - 'null'
...
    Role:
      type: string
      enum:
      - user
      - admin
      - agent
...
    User:
      type: object
      required:
      - id
      - displayName
      - role
      - createdAt
      properties:
        createdAt:
          type: string
          format: date-time
        displayName:
          type: string
        email:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        ownerUserId:
          type:
          - string
          - 'null'
          format: uuid
        role:
          $ref: '#/components/schemas/Role'
...